Extended the drawing procedure to a drawing queue
Process now: if there is a template, load it. Then, if there is a drawing command, add it to the list of drawing tasks. After all that, the things are drawn in order. Todo: if template commands are executed in the beginShow() block, these objects need to be added to the object list